// DXC_Frontend/src/api/models/homestayDto.ts
/**
 * Generated by orval v7.11.2 🍺
 * Do not edit manually.
 * DXC_Core.API
 * OpenAPI spec version: 1.0
 */
import type { HomestayImageDto } from './homestayImageDto'

export interface HomestayDto {
  publicId?: string
  /** @nullable */
  name: string | null
  /** @nullable */
  description?: string | null
  /** @nullable */
  address?: string | null
  /** @nullable */
  phoneNumber?: string | null
  /** @nullable */
  averagePrice?: number | null
  /** @nullable */
  averagePriceCurrency?: string | null
  /** @nullable */
  latitude?: number | null
  /** @nullable */
  longitude?: number | null
  /** @nullable */
  website?: string | null
  /** @nullable */
  linkVitri?: string | null
  thuTu?: number
  isActive?: boolean
  createdAt?: string
  updatedAt?: string
  /** @nullable */
  images?: HomestayImageDto[] | null
}